refactor: move bottom content
